+++ app/Http/Resources/ExportResource.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kami\Cocktail\Http\Resources;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+/**
+ * @mixin \Kami\Cocktail\Models\Export
+ */
+class ExportResource extends JsonResource
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return array<string, mixed>
+     */
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'filename' => $this->filename,
+            'created_at' => $this->created_at,
+            'bar_name' => $this->bar->name,
+            'is_done' => (bool) $this->is_done,
+        ];
+    }
+}