+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v3.12.0
+## New
+- Added `/exports` endpoints
+    - With this endpoint you can now manage recipe exporting for specific bars
+- Added `specific_ingredients` cocktails filter
+    - This will show recipes that always contain specific ingredients
+- Added `public_id` and `slug` to public menu cocktail response
+
+## Fixes
+- Fixed "CocktailParty" scraper
+
 # v3.11.0
 ## New
 - Added POST `/cocktails/{id}/copy` endpoint